pyproject.toml
[tool.poetry]
name = "fastapi_minio_project"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
passlib = "^1.7.4"
jinja2 = "^3.1.4"
icecream = "^2.1.3"
bcrypt = "^4.2.0"
sqlalchemy = "^2.0.32"
python-jose = "^3.3.0"
pydantic-settings = "^2.4.0"
uvicorn = {extras = ["standart"], version = "^0.30.5"}
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
pytest = "^8.3.2"
minio = "^7.2.7"
requests = "^2.32.3"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
